Refactored entire backend into MVC structure
