+++ backend/models/__init__.py
@@ -1,10 +1,8 @@
-# Models
+# Database setup
 from .database import Base
 from .database_session import SessionLocal
+
+# Models
 from .course_prerequisite import CoursePrerequisite
 from .course_corequisite import CourseCorequisite
 from .user_session import UserSession
-
-# Operations
-from . import course_corequisite
-from . import user_session
