+++ backend/routers/__init__.py
@@ -1,0 +1,4 @@
+# Router modules
+from . import user_router
+from . import auth_router
+from . import corequisite_router
