+++ backend/schemas/__init__.py
@@ -1,0 +1,13 @@
+# Pydantic schemas for request/response models
+from .api_models import (
+    SessionPydantic,
+    SessionDeletePydantic,
+    CourseDeletePydantic,
+    updateUser,
+    UserPydantic,
+    UserDeletePydantic,
+    UserCoursePydantic,
+    SubsemesterPydantic,
+    DefaultSemesterSetPydantic,
+    CourseCorequisiteCreate,
+)
