+++ backend/services/__init__.py
@@ -1,0 +1,5 @@
+# Service modules
+from . import corequisite_service
+from . import session_service
+from . import user_service
+from . import auth_service
