+++ backend/services/session_service.py
@@ -1,21 +1,8 @@
-from sqlalchemy import Column, String, DateTime, Integer, ForeignKey
 from sqlalchemy.orm import Session
 from datetime import datetime
-import uuid
 
-from .database import Base
+from models.user_session import UserSession
 
-
-class UserSession(Base):
-    __tablename__ = 'user_session'
-
-    session_id = Column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
-    user_id = Column(Integer, ForeignKey('users.id'), nullable=False)
-    start_time = Column(DateTime, nullable=False)
-    end_time = Column(DateTime, nullable=True)
-
-
-# --- Session operations ---
 
 def create_session(db: Session, user_id: int) -> UserSession:
     """Creates and saves a new user session."""
