+++ backend/services/user_service.py
@@ -1,0 +1,11 @@
+def create_user(user_data: dict):
+    """Placeholder logic to create a new user."""
+    print(f"DATABASE: Creating user '{user_data['username']}'...")
+    # In a real app, you would add the user to the database here.
+    return {"status": "success", "message": f"User {user_data['username']} created."}
+
+def delete_current_user(user_id: int):
+    """Placeholder logic to delete a user."""
+    print(f"DATABASE: Deleting user with ID {user_id}...")
+    # In a real app, you would delete the user from the database here.
+    return {"status": "success", "message": "User deleted."}
